refactor(frontend): implement modular architecture with component separation
