+++ login-app/src/data/barData.ts
@@ -1,0 +1,9 @@
+export const barData = [
+  { name: 'Lunes', Llegaron: 5400, Dejaron: 5400, Presentes: 0 },
+  { name: 'Martes', Llegaron: 6000, Dejaron: 6000, Presentes: 0 },
+  { name: 'Miercoles', Llegaron: 5700, Dejaron: 5700, Presentes: 0 },
+  { name: 'Jueves', Llegaron: 6100, Dejaron: 6100, Presentes: 0 },
+  { name: 'Viernes', Llegaron: 7500, Dejaron: 7500, Presentes: 0 },
+  { name: 'Sabado', Llegaron: 7000, Dejaron: 7000, Presentes: 0 },
+  { name: 'Domingo', Llegaron: 4000, Dejaron: 3700, Presentes: 300 },
+];
