+++ login-app/src/data/lineData.ts
@@ -1,0 +1,14 @@
+export const lineData = [
+  { name: 'Ene', Clientes: 100000 },
+  { name: 'Feb', Clientes: 101250 },
+  { name: 'Mar', Clientes: 90000 },
+  { name: 'Abr', Clientes: 95000 },
+  { name: 'May', Clientes: 105300 },
+  { name: 'Jun', Clientes: 80000 },
+  { name: 'Jul', Clientes: 85000 },
+  { name: 'Ago', Clientes: 87500 },
+  { name: 'Sep', Clientes: 99000 },
+  { name: 'Oct', Clientes: 100500 },
+  { name: 'Nov', Clientes: 102500 },
+  { name: 'Dic', Clientes: 5000 },
+];
