+++ login-app/src/layouts/DashboardLayout.tsx
@@ -1,0 +1,10 @@
+import Sidebar from "../components/Sidebar";
+
+export default function DashboardLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="flex min-h-screen bg-slate-200">
+      <Sidebar />
+      <main className="flex-1 p-6">{children}</main>
+    </div>
+  );
+}
